backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
anthropic==0.39.0
python-dotenv==1.0.1
gunicorn==23.0.0
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2